upgraded libraries, fixed linting errors, fixed all tests
