+++ queue_services/entity-filer/Makefile
@@ -36,7 +36,7 @@ clean-test: ## clean test files
 	rm -fr htmlcov/
 
 build-req: clean ## Upgrade requirements
-	test -f venv/bin/activate || python3.8 -m venv  $(CURRENT_ABS_DIR)/venv ;\
+	test -f venv/bin/activate || python3 -m venv  $(CURRENT_ABS_DIR)/venv ;\
 	. venv/bin/activate ;\
 	pip install pip==20.1.1 ;\
 	pip install -Ur requirements/prod.txt ;\
@@ -45,7 +45,7 @@ build-req: clean ## Upgrade requirements
 	pip install -Ur requirements/bcregistry-libraries.txt
 
 install: clean ## Install python virtrual environment
-	test -f venv/bin/activate || python3.8 -m venv  $(CURRENT_ABS_DIR)/venv ;\
+	test -f venv/bin/activate || python3 -m venv  $(CURRENT_ABS_DIR)/venv ;\
 	. venv/bin/activate ;\
 	pip install pip==20.1.1 ;\
 	pip install -Ur requirements.txt
@@ -64,7 +64,7 @@ lint: ## Linting with pylint
 	. venv/bin/activate && pylint --rcfile=setup.cfg src/$(PROJECT_NAME)
 
 flake8: ## Linting with flake8
-	. venv/bin/activate && flake8 src/$(PROJECT_NAME) tests
+	. venv/bin/activate && flake8 src/$(PROJECT_NAME)
 
 test: ## Unit testing
 	. venv/bin/activate && pytest
