+++ queue_services/entity-filer/requirements.txt
@@ -1,7 +1,7 @@
 Flask==1.1.2
-Jinja2==2.11.2
+Jinja2==2.11.3
 MarkupSafe==1.1.1
-Werkzeug==0.16.1
+Werkzeug==1.0.1
 asyncio-nats-client==0.11.4
 asyncio-nats-streaming==0.4.0
 attr==0.3.1
@@ -9,17 +9,17 @@ attrs==20.3.0
 blinker==1.4
 certifi==2020.12.5
 click==7.1.2
-dpath==1.4.2
+dpath==2.0.1
 itsdangerous==1.1.0
 jsonschema==3.2.0
-protobuf==3.14.0
+protobuf==3.15.6
 pycountry==20.7.3
 pyrsistent==0.17.3
 python-dateutil==2.8.1
 python-dotenv==0.15.0
-sentry-sdk==0.19.4
+sentry-sdk==1.0.0
 six==1.15.0
-urllib3==1.26.2
-git+https://github.com/bcgov/business-schemas.git@2.11.0#egg=registry_schemas
+urllib3==1.26.4
+git+https://github.com/bcgov/business-schemas.git@2.11.3#egg=registry_schemas
 git+https://github.com/bcgov/lear.git#egg=legal_api&subdirectory=legal-api
 git+https://github.com/bcgov/lear.git#egg=entity_queue_common&subdirectory=queue_services/common
