+++ queue_services/entity-filer/requirements/bcregistry-libraries.txt
@@ -1,3 +1,3 @@
-git+https://github.com/bcgov/business-schemas.git@2.11.0#egg=registry_schemas
+git+https://github.com/bcgov/business-schemas.git@2.11.3#egg=registry_schemas
 git+https://github.com/bcgov/lear.git#egg=legal_api&subdirectory=legal-api
 git+https://github.com/bcgov/lear.git#egg=entity_queue_common&subdirectory=queue_services/common
