+++ queue_services/entity-filer/tests/conftest.py
@@ -126,7 +126,7 @@ def account(app):
         print(rv)
 
 
-@pytest.fixture('function')
+@pytest.fixture(scope='function')
 def client_id():
     """Return a unique client_id that can be used in tests."""
     _id = random.SystemRandom().getrandbits(0x58)
